+++ gui/src/renderer/components/Modal.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import ReactDOM from 'react-dom';
 import { Component, Styles, Text, View } from 'reactxp';
 import { colors } from '../../config.json';
+import { Scheduler } from '../../shared/scheduler';
 import ImageView from './ImageView';
 
 const MODAL_CONTAINER_ID = 'modalContainer';
@@ -10,8 +11,9 @@ const styles = {
   modalAlertBackground: Styles.createViewStyle({
     flex: 1,
     justifyContent: 'center',
-    paddingLeft: 14,
-    paddingRight: 14,
+    paddingHorizontal: 14,
+    paddingTop: 26,
+    paddingBottom: 14,
   }),
   modalAlert: Styles.createViewStyle({
     backgroundColor: colors.darkBlue,
@@ -20,15 +22,15 @@ const styles = {
   }),
   modalAlertIcon: Styles.createViewStyle({
     alignItems: 'center',
-    marginBottom: 12,
-    marginTop: 4,
+    marginTop: 8,
   }),
   modalAlertMessage: Styles.createTextStyle({
     fontFamily: 'Open Sans',
-    fontSize: 16,
+    fontSize: 14,
     fontWeight: '500',
     lineHeight: 20,
     color: colors.white80,
+    marginTop: 16,
   }),
   modalAlertButtonContainer: Styles.createViewStyle({
     marginTop: 16,
@@ -107,18 +109,27 @@ interface IModalAlertProps {
 export class ModalAlert extends Component<IModalAlertProps> {
   private element = document.createElement('div');
   private modalContainer?: Element;
+  private appendScheduler = new Scheduler();
 
   public componentDidMount() {
     const modalContainer = document.getElementById(MODAL_CONTAINER_ID);
     if (modalContainer) {
       this.modalContainer = modalContainer;
-      modalContainer.appendChild(this.element);
+
+      // Mounting the container element immediately results in a graphical issue with the dialog
+      // first rendering with the wrong proportions and then changing to the correct proportions.
+      // Postponing it to the next event cycle solves this issue.
+      this.appendScheduler.schedule(() => {
+        modalContainer.appendChild(this.element);
+      });
     } else {
       throw Error('Modal container not found when mounting modal');
     }
   }
 
   public componentWillUnmount() {
+    this.appendScheduler.cancel();
+
     if (this.modalContainer) {
       this.modalContainer.removeChild(this.element);
     }
@@ -136,9 +147,7 @@ export class ModalAlert extends Component<IModalAlertProps> {
             {this.props.type && (
               <View style={styles.modalAlertIcon}>{this.renderTypeIcon(this.props.type)}</View>
             )}
-            {this.props.message && (
-              <Text style={styles.modalAlertMessage}>{this.props.message}</Text>
-            )}
+            {this.props.message && <ModalMessage>{this.props.message}</ModalMessage>}
             {this.props.children}
             {this.props.buttons.map((button, index) => (
               <View key={index} style={styles.modalAlertButtonContainer}>
@@ -166,4 +175,12 @@ export class ModalAlert extends Component<IModalAlertProps> {
     }
     return <ImageView height={44} width={44} source={source} tintColor={color} />;
   }
+}
+
+interface IModalMessageProps {
+  children?: string;
+}
+
+export function ModalMessage(props: IModalMessageProps) {
+  return <Text style={styles.modalAlertMessage}>{props.children}</Text>;
 }